+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaMirrorMaker2ClusterTest.java
@@ -198,7 +198,7 @@ public class KafkaMirrorMaker2ClusterTest {
     protected List<EnvVar> getExpectedEnvVars() {
         List<EnvVar> expected = new ArrayList<>();
         expected.add(new EnvVarBuilder().withName(KafkaMirrorMaker2Cluster.ENV_VAR_KAFKA_CONNECT_CONFIGURATION).withValue(expectedConfiguration.asPairs()).build());
-        expected.add(new EnvVarBuilder().withName(KafkaMirrorMaker2Cluster.ENV_VAR_KAFKA_CONNECT_METRICS_ENABLED).withValue(String.valueOf(true)).build());
+        expected.add(new EnvVarBuilder().withName(KafkaMirrorMaker2Cluster.ENV_VAR_KAFKA_CONNECT_JMX_EXPORTER_ENABLED).withValue(String.valueOf(true)).build());
         expected.add(new EnvVarBuilder().withName(KafkaMirrorMaker2Cluster.ENV_VAR_KAFKA_CONNECT_BOOTSTRAP_SERVERS).withValue(bootstrapServers).build());
         expected.add(new EnvVarBuilder().withName(KafkaMirrorMaker2Cluster.ENV_VAR_STRIMZI_KAFKA_GC_LOG_ENABLED).withValue(Boolean.toString(JvmOptions.DEFAULT_GC_LOGGING_ENABLED)).build());
         expected.add(new EnvVarBuilder().withName(AbstractModel.ENV_VAR_KAFKA_HEAP_OPTS).withValue(kafkaHeapOpts).build());
@@ -2236,10 +2236,25 @@ public class KafkaMirrorMaker2ClusterTest {
     @ParallelTest
     public void testMetricsParsingNoMetrics() {
         KafkaMirrorMaker2Cluster kmm = KafkaMirrorMaker2Cluster.fromCrd(Reconciliation.DUMMY_RECONCILIATION, this.resource, VERSIONS, SHARED_ENV_PROVIDER);
+        assertThat(kmm.metrics(), is(nullValue()));
+    }
 
-        assertThat(kmm.metrics().isEnabled(), is(false));
-        assertThat(kmm.metrics().getConfigMapName(), is(nullValue()));
-        assertThat(kmm.metrics().getConfigMapKey(), is(nullValue()));
+    @ParallelTest
+    public void testStrimziMetricsReporterConfig() {
+        KafkaMirrorMaker2 resourceWithMetrics = new KafkaMirrorMaker2Builder(resource)
+            .editSpec()
+                .withNewStrimziMetricsReporterConfig()
+                    .withNewValues()
+                        .withAllowList(List.of("kafka_log.*", "kafka_network.*"))
+                    .endValues()
+                .endStrimziMetricsReporterConfig()
+            .endSpec()
+            .build();
+        
+        InvalidResourceException ex = assertThrows(InvalidResourceException.class,
+            () -> KafkaMirrorMaker2Cluster.fromCrd(Reconciliation.DUMMY_RECONCILIATION, resourceWithMetrics, VERSIONS, SHARED_ENV_PROVIDER));
+
+        assertThat(ex.getMessage(), is("The Strimzi Metrics Reporter is not supported for this component"));
     }
 
     @ParallelTest